+++ tests/fixtures/sydtest/cabal/simple/src/Lib.hs
@@ -1,0 +1,8 @@
+module Lib
+  ( someFunc
+  )
+where
+
+someFunc :: IO ()
+someFunc = putStrLn "someFunc"
+